infra/requirements.txt
aws-cdk.core>=1.160.0
aws-cdk.assertions>=1.160.0
cdk-chalice==0.9.0
